Implementacion de auth, token, detalle de survey y ultimos detalles

+++ app/src/main/java/com/antonelli/nimble/entity/TokenModel.kt
@@ -3,7 +3,7 @@ package com.antonelli.nimble.entity
 data class TokenModel(
     val access_token: String? = null,
     val token_type: String? = null,
-    val expires_in: Int? = null,
+    val expires_in: Double? = null,
     val refresh_token: String? = null,
-    val created_at: Int? = null
+    val created_at: Double? = null
 )
